Initial regex impl working
This is following the VERY minimal example given by Russ Cox in his
article so it will only match against the full length of the string
rather than substring matches and it only reports whether or not it
matched.
Next up is making this something more usable and looking at the caching
strategies covered in the article.