Adding some simple optimisations for the VM regex engine
Once we've compiled to the VM op-code form there are a couple of simple
optimisations we can make to the resulting program. The first is to
inline any repeated jumps or splits to jumps rather than leaving them as
additional instructions for the VM to execute. The second is to note
that a jump to Match can also be inlined as that will end the current VM
thread. Finally (for now) we can also search the optimised program for
any instructions that are no longer reachable due to inlining and remove
them so we can save memory on both the stored program and clist / nlist
caches. I'm defining an instruction as unreachable if nothing jumps to
it and normal VM thread execution would not advance to it (so the
previous instruction is not a comparison op). Removing instructions can
(and often will) invalidate other Jump and Split instructions in the
program so we need to rewrite those to correct their offsets as well
once we have removed the unreachable instructions.