Initial wiring in of the exec mode to the editor
This is very much MVP at this stage. Buffer now implements the trait required
for running an exec program with the intention that it correctly tracks
edits so they can be undone if needed. That side of things isn't working
correctly as whitespace changes (I think?) seem to be getting lost and
it appears to be possible to undo to a point that the edit log is
invalid, resulting in a panic...

So, still a lot to get done! But on the plus side, the in-place editing
all works as intended and slurping output to a mini-buffer seems to do
the right thing as well.