Using DotExpression in exec and fixing match location semantics
This commit is a bit of a mess as I ended up heading down a rabbit hole
trying to unpick why '.*' wasn't behaving in the way that I expected
when looping over lines using x/re/ in exec. The initial short answer
was that I was using inclusive range semantics for Match.loc() which
then doesn't support expressing a match of a 0-length string which is
needed for '.*' (and '*' in general). Moving the match location
semantics over to not include the endpoint broke A LOT of things (as you
would expect) and unpicking all of that resulted in this commit.

One outstanding issue is that I have ended up having to add a check that
the position of Save(0) is >= the initial value of sp in the regex VM
and I really can't get my head round how it is ended up either unset (so
defaulting to 0) or explicitly being set to 0. Hopefully some more
coffee and a second look should sort it out but before jumping down THAT
rabbit hole I'm pushing what I have so far.