Fixing bug with checking final match state
The issue with needing to clear stale threads was down to how I was
handling breaking from the main loop if the thread list we were building
ended up empty. The update to the 'n' variable used to then check the final
thread list was being skipped meaning that we checked up to the length
of the _previous_ list, ending up pulling in old states. The fix is
simply to move the update to n above the conditional break out of the
loop (which then means we don't need to explicitly zero out old thread
submatches).