initial read only fsys now working
This marks the initial integration of the acme style filesystem
interface into the rest of the editor. For now all you are able to do is
read the various Buffer control files, writing is not supported. Next up
is hooking up writing to each of the various control files and then once
that's working I need to have a think about how to set up and handle the
events file...