fix: improving the performance of tree-sitter highlighting
The initial implementation of tree-sitter based highlighting was
knowingly simple and naive and worked well enough, but had a number
of issues. The main one (other than incorrect highlighting which has
already been addressed) was that perfomance took a massive hit once you
started working with larger files. The implementation now is a big
improvement but its still not ideal and will need more work in the
future.