fix: scrolling multiple views of the same buffer
Previously scrolling a second view of the active buffer would jump the
viewport to the current cursor position before scrolling (which was
really annoying). This changes things to allow for independent scrolling
of each view, only updating the cursor position when focus moves to that
view. This probably still needs a bit of work as the views are not
tracking their desired cursor position so that it is maintained when
they are focused again, causing the cursor to move if you switch back
and forth.