replacing the virtual text banner with in buffer splash message
The welcome banner (using virtual text) has been present since the very
early development of ad but it doesn't actually fit the stated goal of
all in-buffer text being something you can manipulate and load/execute
(in addition to behaving strangely when the banner is shown and multiple
windows are present). To rectify this, I've moved to having a config
flag that allows for a canned splash message to be shown on startup in
the default scratch buffer if no files are open. This is then normal
in-buffer text that can be interacted with as the user expects.