Documenting why Range is closed rather than half-open as you would expect
After spending an afternoon trying to rework Range to convert it to a
traditional half-open form I have rediscovered why it is this way in the
first place. Namely, we need to be able to manipulate each end of the
range as an independent Cursor and recombine arbitrary Cursors into new
Ranges. While it is possible to get this to work with a half-open
interval you end up with a fair amount of book keeping to track and you
lose the ability to combine arbitrary Cursors together into Ranges as
you have no way of knowing whether or not the end cursor should be
incremented to create a half-open interval or if the caller has already
handled that for you.

Yes it would be possible to rework this but it is a nasty thing to
unpick and the benefit in the end is fairly minimal given that the
current implementation functions correctly and has clear semantics:
  - ranges are inclusive
  - slicing requires a half-open interval so increment the end
  - extending a range by moving the "active cursor" to a new index is
    well defined regardless of which end of the range is active