Correcting highlighting for selected ranges
The logic for applying highlighting for dot, load and exec ranges had an
off by one error and stopped a character short of the correct range.
Normally you don't notice this that ofter as the cursor sits at that
position and takes priority over showing the background color. I've been
meaning to fix this for a while but I kept forgetting to look at it.