Issue #98 - initial impl for resizing window and columns
So far this adds two new commands "resize-col" and "resize-win" that
both take a single argument of a signed integer and resize the active
column or window respectively by that delta (in terms of character
cells). Resizing clamps so that the column/window being shrunk is at
least 5 cells wide/tall but currently any call to update_size will
result in the user customised layout being trashed. Fixing that is up
next.