Adding support for fsys in scenario tests
The usage of a OnceCell for the editor config means that these tests
need to run serially in order to be able to ensure that they run with
the correct config and that they don't stomp on one another.

Until now, the use of global state for the config hasn't caused any
issues for the editor itself but this has highlighted that it can result
in some subtle and infuriating bugs within tests where multiple copies
of the editor are being run and there is a desire to use something other
than the default config. (Who could have seen this one coming...)

I probably _do_ need to refactor the config to live in the editor itself
and instead place it in an Arc but that's a problem for future me to
figure out.