Factoring out logic for opening initial files
Pretty much since the beginning of the project, the logic for looping
over the command line args and opening any initially requested files has
lived in main.rs. Given that config errors (and now plumbing rules errors)
are handled and opened in virtual buffers as part of this, I had to
duplicate this logic in the new scenario tests when they were first
written, which wasn't ideal. Now this has been factored out into a new
helper method to construct the editor from _results_ for the config and
plumbing rules so the handling of how those error messages are displayed
to the user can be reused.