Add reference editing tests
The README for the reference-tests directory has links to where the test
data and reference has come from (https://github.com/josephg/editing-traces/)
as well a new script for bisecting a failing edit sequence to pin down
where things diverge.

Running this has shown that there was a bug around calculating the raw
byte position for GapBuffer operations which I have a fix for in the
following commit.