Fix bug with finding raw byte indices in GapBuffer
When computing the raw byte index for a GapBuffer insert or delete when
the logical index in question ends up being the start of the current
gap, we were incorrectly returning the start of the gap rather than the
end as the index.

In practice this only seems to have been an issue inside of
GapBuffer::remove_char, but fixing it then broke the calculation for
Slice offsets which also needed fixing.