Fix handling for terminal mouse events (wide chars)
It looks like this has probably always been broken: the x coordinate
received from the terminal when using SGR (1006) mouse coordinates is in
terms of terminal cells, not characters. This means that we need to
accound for wide utf8 characters when we map a click location into a new
Dot for the Buffer containing the mouse click.

The change in src/buffer/mod.rs is the one line fix for this but it
draws attention to the fact that there is a fair amount of logic at the
moment that is specific to the terminal UI which exists throughout the
main editor logic. At some point I'm going to need to move things like
this out to the UserInterface trait if I do end up writing a GUI front
end.