Improve GapBuffer performance
We can improve the performance of the GapBuffer by caching the results
of calls to `offset_char_to_byte` until the next buffer modification
when they potentially become invalidated.

We can also improve the performance of treesitter parsing from GapBuffer
state by not needlessly mapping back and forth between character and
byte offsets in a _very_ inefficient way...

Under the existing criterion benchmarks this looks to result in roughly
an 18% speedup for syntax highlighting on my machine.