Prevent crash in keybinding parsing for Config
As reported in #158, the simple parser for custom keybindings was
triggering a panic for some keybindings that looked like they should
have been valid. The parser was only supporting a subset of the keys
that ad can handle in the form of simple string templates, and was
intending to return an error if it didn't recognise the key template
it had been given. This error was being dropped due to a flat_map though
and there was no guard against attempting to construct a Trie from an
empty input which was where the panic was coming from.

This commit extends the parser to cover inputs other than arrows and
mouse clicks and provides some tests for the behaviour of the parser.