Fix behaviour of send_keys
In adding the scenario tests in this commit, I found that the behaviour
of "send_keys" was incorrect in that it required the key sequence to be
interpreted as a single Trie key rather than a sequence of inputs that
were then handled in the same way as normal inputs from the user. The
fix is simple and the new scenario tests validate the behaviour.