Fix calculation of tree-sitter edits
I was incorrectly computing the byte offset for the `old_end_byte`
parameter of tree sitter's InputEdit struct using the newly modified
buffer. Almost all of the time this was fine, but very occasionally I
was seeing the syntax highlighting for a file become broken and drift
from the content of the buffer. Interestingly, it isn't a case of the
highlighting _completely_ breaking, or anything crashing, just that the
highlighting spans end up misaligned compared to the actual buffer
content.

I'm not 100% convinced that this is the only thing that needs fixing in
this area, as I've not actually been able to write a failing test to
reproduce the bug. But, this is definitely a change that needed to be
made anyway so I'm putting it in place now and will likely need to
revisit things at a later date.