Enforce POSIX line semantics
This used to be a feature of ad but I removed it a while back as the
implementation I had was...funky. The issue is that I want to enforce
POSIX line semantics (all lines are _terminated_ with a newline, not
_delimited_) but I also wanted to sort out the rendering logic in order to
not have this display the final newline as a character at the end of the
buffer that you can delete. Not doing this last part means that buffers
end up with what looks like a trailing blank line which the user is
sorely tempted to remove as it just looks odd.

What I was doing before was popping off the trailing newline on file
load and then putting it back before saving (which was disgusting).
_Not_ having the final newline be a character you can delete is really
weird and forces a bunch of special casing in the rest of the editor
that I really don't like. The approach I'm taking here is to indicate in
the UI when you don't have a trailing newline and then insert one prior
to saving so that if you do accidentally (or deliberately) delete the
trailing newline it gets put back.